tax-ai-social/requirements.txt
flask>=3.0.0
requests>=2.31.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
